Moved single plant component from full plant gallery (Gallery) into it's own component (GalleryItem).
