+++ app/components/index.js
@@ -2,5 +2,6 @@ export { default as Root } from './Root';
 export { default as AddPlant } from './AddPlant';
 export { default as Dashboard } from './Dashboard';
 export { default as EditPlant } from './EditPlant';
-export { default as Plants } from './Plants';
+export { default as Gallery } from './Gallery';
 export { default as SinglePlant } from './SinglePlant';
+export { default as GalleryItem } from './GalleryItem';
