+++ app/components/root.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Route, Switch, NavLink } from 'react-router-dom';
-import { AddPlant, Plants, SinglePlant, EditPlant, Dashboard } from './';
+import { AddPlant, Gallery, SinglePlant, EditPlant, Dashboard } from '.';
 import { withStyles } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import CssBaseline from '@material-ui/core/CssBaseline';
@@ -13,11 +13,14 @@ const Root = () => {
       <CssBaseline />
       <div>
         <nav>
-          <NavLink to={'/dashboard'} className="navlink">
-            DASHBOARD
+          <NavLink to={'/gallery'} className="navlink">
+            GALLERY
           </NavLink>
-          <NavLink to={'/plants'} className="navlink">
-            COLLECTION
+          <NavLink to={'/dashboard'} className="navlink">
+            SCHEDULE
+          </NavLink>
+          <NavLink to={'/dashboard'} className="navlink">
+            TRENDS
           </NavLink>
         </nav>
         <main>
@@ -29,9 +32,9 @@ const Root = () => {
               component={EditPlant}
             />
             <Route path="/plants/:plantId" component={SinglePlant} />
-            <Route path="/plants" component={Plants} />
+            <Route path="/gallery" component={Gallery} />
             <Route path="/dashboard" component={Dashboard} />
-            <Route path="/" component={Dashboard} />
+            <Route path="/" component={Gallery} />
           </Switch>
         </main>
       </div>
